requirements.txt
fastapi
ngrok
python-dotenv
sqlalchemy
psycopg2-binary
pg8000
alembic

